docker_submission/scripts/generativeMRI/requirements.txt
pytorch==2.2.0
pytorch-cuda==12.1
torchaudio==2.2.0
torchio==0.19.5
torchmetrics==1.2.1
torchvision==0.17.0
monai==1.3.0
monai-generative==0.2.3
pytorch-lightning==2.1.3
scikit-image==0.22.0
scikit-learn==1.4.1
scipy==1.11.4
simpleitk==2.2.1
pillow==10.2.0
nibabel==5.2.0
seaborn==0.12.2
matplotlib==3.5.3
tqdm==4.65.0
numpy==1.26.3
imageio==2.33.1
lpips
tensorboard
wandb==0.16.3
raster_geometry==0.1.4.2

